Add macro for disabling Python dist metadata dependency generator
This macro is borrowed from Fedora to be able to optionally disable
the new dependency generator in RPM 4.15 that generates pythonX.Ydist(M)
requires.

This is mainly intended for cases where the metadata is flat out wrong/broken
or if the metadata does not map to the packaging granularity (that is,
a Python package has been split into several subpackages and dependencies
are split along with them).