Add "export" PYTHONPATH for the correct usage inside of the installation directory
PYTHONPATH has been set at the moment but has not been exported. Therefore, some python tests can not find the installation directory with built modules by the package ( #124 ). In my case, I can not build python-zstandard because of this issue. The "export" has got the effect of the usage of the correct PYTHONPATH in pytest.