Pass `--ignore-installed` to `pip install` in %pyproject_install
Packages that put the source next to pyproject.toml (instead of e.g.
into the src/ directory) will usually have a $name.egg-info after build.
For each $name.egg-info directory that is found in sys.path, pip
considers package $name as installed. The `-m` argument to `python`
causes the current directory to be added to sys.path. Pip >= 20.1
removes that directory [1]. Leap 15.3 and 15.4 use 20.0.2, so pip thinks
that the package is already installed if $name.egg-info is in the
directory where %pyproject_install is run.

This is fixed by passing `--ignore-installed` to pip. This makes sense
anyway. We pass `--root` to pip. Even in this case, pip will search for
installed packages as if `--root` was not passed. The build root we pass
is empty and it wouldn’t search inside it anyway, so it makes sense to
tell pip to not search for installed packages at all.

See https://bugzilla.opensuse.org/show_bug.cgi?id=1201041.

[1] https://github.com/pypa/pip/pull/7955